refactor(security-utils): separate api error sanitizer
