+++ pages/options/AuthManager.js
@@ -1,6 +1,7 @@
 /* global chrome */
 import Logger from '../../scripts/utils/Logger.js';
-import { sanitizeApiError, createSafeIcon } from '../../scripts/utils/securityUtils.js';
+import { createSafeIcon } from '../../scripts/utils/securityUtils.js';
+import { sanitizeApiError } from '../../scripts/utils/ApiErrorSanitizer.js';
 import { ErrorHandler } from '../../scripts/utils/ErrorHandler.js';
 import { UI_MESSAGES } from '../../scripts/config/shared/messages.js';
 import { UI_ICONS } from '../../scripts/config/icons.js';
