+++ pages/options/MigrationScanner.js
@@ -30,7 +30,7 @@
 
 import Logger from '../../scripts/utils/Logger.js';
 import { RUNTIME_ACTIONS } from '../../scripts/config/shared/runtimeActions.js';
-import { sanitizeApiError } from '../../scripts/utils/securityUtils.js';
+import { sanitizeApiError } from '../../scripts/utils/ApiErrorSanitizer.js';
 import { ErrorHandler } from '../../scripts/utils/ErrorHandler.js';
 
 /**
