+++ pages/options/StorageManager.js
@@ -8,11 +8,11 @@
 
 import Logger from '../../scripts/utils/Logger.js';
 import {
-  sanitizeApiError,
   validateSafeSvg,
   separateIconAndText,
   createSafeIcon,
 } from '../../scripts/utils/securityUtils.js';
+import { sanitizeApiError } from '../../scripts/utils/ApiErrorSanitizer.js';
 import { ErrorHandler } from '../../scripts/utils/ErrorHandler.js';
 import { UI_ICONS } from '../../scripts/config/icons.js';
 import { UI_MESSAGES } from '../../scripts/config/shared/messages.js';
