+++ pages/popup/popupActions.js
@@ -6,7 +6,8 @@
 
 /* global chrome */
 
-import { isValidNotionUrl, sanitizeApiError } from '../../scripts/utils/securityUtils.js';
+import { isValidNotionUrl } from '../../scripts/utils/securityUtils.js';
+import { sanitizeApiError } from '../../scripts/utils/ApiErrorSanitizer.js';
 import { sanitizeUrlForLogging } from '../../scripts/utils/LogSanitizer.js';
 import Logger from '../../scripts/utils/Logger.js';
 import { AuthMode } from '../../scripts/config/extension/authMode.js';
