+++ scripts/background/handlers/handlerGuard.js
@@ -8,11 +8,8 @@
 
 /* global Logger */
 
-import {
-  isValidUrl,
-  sanitizeApiError,
-  validateInternalRequest,
-} from '../../utils/securityUtils.js';
+import { isValidUrl, validateInternalRequest } from '../../utils/securityUtils.js';
+import { sanitizeApiError } from '../../utils/ApiErrorSanitizer.js';
 import { sanitizeUrlForLogging } from '../../utils/LogSanitizer.js';
 import { ErrorHandler } from '../../utils/ErrorHandler.js';
 import { ERROR_MESSAGES } from '../../config/shared/messages.js';
