+++ scripts/background/handlers/saveHandlers.js
@@ -13,8 +13,8 @@ import {
   validateInternalRequest,
   validateContentScriptRequest,
   isValidNotionUrl,
-  sanitizeApiError,
 } from '../../utils/securityUtils.js';
+import { sanitizeApiError } from '../../utils/ApiErrorSanitizer.js';
 import { sanitizeUrlForLogging } from '../../utils/LogSanitizer.js';
 import { buildHighlightBlocks } from '../utils/BlockBuilder.js';
 import { parseArgsToContext } from '../../utils/Logger.js';
