+++ scripts/background/services/NotionService.js
@@ -16,7 +16,7 @@ import { ERROR_MESSAGES, CONTENT_QUALITY, HIGHLIGHT_ERROR_CODES } from '../../co
 import { NOTION_API } from '../../config/extension/notionApi.js';
 import { AuthMode } from '../../config/extension/authMode.js';
 // 導入安全工具
-import { sanitizeApiError } from '../../utils/securityUtils.js';
+import { sanitizeApiError } from '../../utils/ApiErrorSanitizer.js';
 // 導入暫存圖片 URL 偵測（用於 page cover 防護）
 // 從獨立模組 import，避免 rollup 被迫保留 imageUtils 整個物件含所有函數
 import { isTemporaryImageUrl } from '../../utils/temporaryImageUrl.js';
