+++ scripts/utils/notionAuth.js
@@ -4,7 +4,7 @@ import { AuthMode } from '../config/extension/authMode.js';
 import { NOTION_OAUTH } from '../config/extension/notionAuth.js';
 import { BUILD_ENV } from '../config/env/index.js';
 import { RUNTIME_ACTIONS } from '../config/shared/runtimeActions.js';
-import { sanitizeApiError } from './securityUtils.js';
+import { sanitizeApiError } from './ApiErrorSanitizer.js';
 import { ERROR_MESSAGES } from '../config/shared/messages.js';
 
 const AUTH_EPOCH_KEY = 'notionAuthEpoch';
