+++ tests/unit/background/handlers/actionHandlers.edge-cases.test.js
@@ -69,9 +69,12 @@ jest.mock('../../../../scripts/utils/securityUtils.js', () => ({
     }
     return null;
   }),
-  sanitizeApiError: jest.fn(err => (err instanceof Error ? err.message : String(err))),
   isValidNotionUrl: jest.fn(() => true),
   sanitizeUrlForLogging: jest.fn(url => url),
+}));
+
+jest.mock('../../../../scripts/utils/ApiErrorSanitizer.js', () => ({
+  sanitizeApiError: jest.fn(err => (err instanceof Error ? err.message : String(err))),
 }));
 
 // 引入測試所需模組
